+++ CHANGELOG.md
@@ -4,6 +4,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.14.0] - 2021-03-22
+
+### Added
+
+- Expose lifecycle hook events. This feature allows sending VWO data to third party integrations.
+
+### Changed
+
+- Introduce `integrations` key in `VWO` class constructor's argument to enable receiving hooks for the third party integrations.
+
+```php
+$config = [
+    'settingsFile' => $settingsFile,
+    'integrations' => [
+        'callback' => function ($properties) {
+            // $properties will contain all the required VWO specific information
+        }
+    ]
+];
+
+$vwoClientInstance = new VWO($config);
+```
+
 ## [1.13.1] - 2021-03-18
 
 ### Changed